ai-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
clickhouse-connect==0.7.19
prophet==1.1.6
pandas==2.2.2
numpy==1.26.4
python-dotenv==1.0.1
